Moved the check_string_dictionary function to dnac.py
